[CODE UNIFICATION] Use the PrebidRenderingTargeting class instead of OriginalTargetingProtocol (#557)
* feat: confirm PrebidRenederingTargeting to OriginalTargetingProtocol

* feat: rename PrebidRenderingTargeting to Targeting & use Targeting

* fix: fix test apps

* feat(tests): update tests

* chore: move files to respective folder

* fix: fix naming
