+++ PrebidMobile/PrebidMobileRendering/Networking/Parameters/PBMParameterBuilderService.m
@@ -55,7 +55,7 @@
                                    sdkConfiguration:Prebid.shared
                                          sdkVersion:[PBMFunctions sdkVersion]
                               pbmUserConsentManager:PBMUserConsentDataManager.shared
-                                          targeting:PrebidRenderingTargeting.shared
+                                          targeting:Targeting.shared
                              extraParameterBuilders:extraParameterBuilders];
 }
 
@@ -70,7 +70,7 @@
                                                                     sdkConfiguration:(nonnull Prebid *)sdkConfiguration
                                                                           sdkVersion:(nonnull NSString *)sdkVersion
                                                                pbmUserConsentManager:(nonnull PBMUserConsentDataManager *) pbmUserConsentManager
-                                                                           targeting:(nonnull PrebidRenderingTargeting *)targeting
+                                                                           targeting:(nonnull Targeting *)targeting
                                                               extraParameterBuilders:(nullable NSArray<id<PBMParameterBuilder> > *)extraParameterBuilders{
   
     PBMORTBBidRequest *bidRequest = [PBMParameterBuilderService createORTBBidRequestWithTargeting:targeting];
@@ -100,7 +100,7 @@
     return [PBMORTBParameterBuilder buildOpenRTBFor:bidRequest];
 }
 
-+ (nonnull PBMORTBBidRequest *)createORTBBidRequestWithTargeting:(nonnull PrebidRenderingTargeting *)targeting {
++ (nonnull PBMORTBBidRequest *)createORTBBidRequestWithTargeting:(nonnull Targeting *)targeting {
     PBMORTBBidRequest *bidRequest = [PBMORTBBidRequest new];
     
     bidRequest.user.yob = targeting.userAge > 0 ?
