+++ PrebidMobile/PrebidMobileRendering/Networking/Parameters/PBMSKAdNetworksParameterBuilder.m
@@ -30,7 +30,7 @@
 @property (nonatomic, class, readonly) NSString *SKAdNetworkIdentifierKey;
 
 @property (nonatomic, strong, readonly) id<PBMBundleProtocol> bundle;
-@property (nonatomic, strong, readonly) PrebidRenderingTargeting *targeting;
+@property (nonatomic, strong, readonly) Targeting *targeting;
 @end
 
 #pragma mark - Implementation
@@ -50,7 +50,7 @@
 
 #pragma mark - Initialization
 
-- (nonnull instancetype)initWithBundle:(id<PBMBundleProtocol>)bundle targeting:(PrebidRenderingTargeting *)targeting {
+- (nonnull instancetype)initWithBundle:(id<PBMBundleProtocol>)bundle targeting:(Targeting *)targeting {
     if (!(self = [super init])) {
         return nil;
     }
