+++ PrebidMobile/PrebidMobileRendering/Prebid/Integrations/MediationAPI/MediationBannerAdUnit.swift
@@ -146,7 +146,7 @@ public class MediationBannerAdUnit : NSObject {
             
             self.fetchDemand(connection: PBMServerConnection.shared,
                              sdkConfiguration: Prebid.shared,
-                             targeting: PrebidRenderingTargeting.shared,
+                             targeting: Targeting.shared,
                              completion: completion)
         })
     }
@@ -155,7 +155,7 @@ public class MediationBannerAdUnit : NSObject {
         
         fetchDemand(connection: PBMServerConnection.shared,
                     sdkConfiguration: Prebid.shared,
-                    targeting: PrebidRenderingTargeting.shared,
+                    targeting: Targeting.shared,
                     completion: completion)
     }
     
@@ -175,7 +175,7 @@ public class MediationBannerAdUnit : NSObject {
     // NOTE: do not use `private` to expose this method to unit tests
     func fetchDemand(connection: PBMServerConnectionProtocol,
                      sdkConfiguration: Prebid,
-                     targeting: PrebidRenderingTargeting,
+                     targeting: Targeting,
                      completion: ((FetchDemandResult)->Void)?) {
         guard bidRequester == nil else {
             // Request in progress
