+++ PrebidMobile/PrebidMobileRendering/Prebid/Integrations/MediationAPI/MediationBaseInterstitialAdUnit.swift
@@ -40,7 +40,7 @@ public class MediationBaseInterstitialAdUnit : NSObject {
     public func fetchDemand(completion: ((FetchDemandResult)->Void)?) {
         fetchDemand(connection: PBMServerConnection.shared,
                     sdkConfiguration: Prebid.shared,
-                    targeting: PrebidRenderingTargeting.shared,
+                    targeting: Targeting.shared,
                     completion: completion)
     }
     
@@ -103,7 +103,7 @@ public class MediationBaseInterstitialAdUnit : NSObject {
     // NOTE: do not use `private` to expose this method to unit tests
     func fetchDemand(connection: PBMServerConnectionProtocol,
                      sdkConfiguration: Prebid,
-                     targeting: PrebidRenderingTargeting,
+                     targeting: Targeting,
                      completion: ((FetchDemandResult)->Void)?) {
         guard bidRequester == nil else {
             // Request in progress
