+++ PrebidMobileTests/RenderingTests/TestExtensions/PBMBasicParameterBuilder+pbmTestExtension.h
@@ -19,7 +19,7 @@
 
 @property (nonatomic, strong, nullable, readwrite) PBMAdConfiguration *adConfiguration;
 @property (nonatomic, strong, nullable, readwrite) Prebid *sdkConfiguration;
-@property (nonatomic, strong, nullable, readwrite) PrebidRenderingTargeting *targeting;
+@property (nonatomic, strong, nullable, readwrite) Targeting *targeting;
 @property (nonatomic, copy, nullable, readwrite) NSString *sdkVersion;
 
 @end
