+++ PrebidMobileTests/RenderingTests/TestExtensions/Prebid/MediationInterstitialAdUnit+TestExtension.h
@@ -19,6 +19,6 @@
 - (void)fetchDemandWithObject:(NSObject *)adObject
                    connection:(id<PBMServerConnectionProtocol>)connection
              sdkConfiguration:(Prebid *)sdkConfiguration
-                    targeting:(PrebidRenderingTargeting *)targeting
+                    targeting:(Targeting *)targeting
                    completion:(void (^)(FetchDemandResult))completion;
 @end
