+++ PrebidMobileTests/RenderingTests/Tests/ParameterBuilderTests/PBMAppInfoParameterBuilderTest.swift
@@ -22,13 +22,13 @@ class PBMAppInfoParameterBuilderTest: XCTestCase {
     var bidRequest: PBMORTBBidRequest!
     var mockBundle: MockBundle!
     var builder: PBMAppInfoParameterBuilder!
-    var targeting: PrebidRenderingTargeting!
+    var targeting: Targeting!
     
     override func setUp() {
         super.setUp()
         bidRequest = PBMORTBBidRequest()
         mockBundle = MockBundle()
-        targeting = PrebidRenderingTargeting.shared
+        targeting = Targeting.shared
         targeting.publisherName = publisherName
         
         builder = PBMAppInfoParameterBuilder(bundle: mockBundle, targeting: targeting)
